portfolio/requirements.txt
Flask==2
python-decouple==3.4
requests==2.25
